center aligned PWM

+++ Core/Inc/pwm.h
@@ -6,25 +6,25 @@ extern "C" {
 #endif
 
 struct Reference {
-    double angle;
-    double magnitude; 
+    float angle;
+    float magnitude; 
 };
 
 // initialize all your state vectors with CCR value
 struct StateVectors {
-    int V_100[3];
-    int V_110[3];
-    int V_010[3];
-    int V_011[3];
-    int V_001[3];
-    int V_101[3];
+    float V_100[3];
+    float V_110[3];
+    float V_010[3];
+    float V_011[3];
+    float V_001[3];
+    float V_101[3];
 };
 
-void initialize_reference(struct Reference* vec, double angle, double magnitude);
-void initialize_array(struct StateVectors* svpwm_obj, int CCR);
-double convert_rpm_to_angle(double rpm, double period);
+void initialize_reference(struct Reference* vec, float angle, float magnitude);
+void initialize_array(struct StateVectors* svpwm_obj, float CCR);
+float convert_rpm_to_angle(float rpm, float period);
 struct Vec3 update_CCR(struct StateVectors* obj, struct Reference* vector);
-struct Vec3 update_CCR_new(double duty_cycle, double pwm_period, struct Reference* vector);
+struct Vec3 update_CCR_new(float duty_cycle, float pwm_period, struct Reference* vector);
 
 #ifdef __cplusplus
 }
